[유소희]파일 시퀀스 타입 변경(Long -> int)

+++ src/main/java/com/loanscrefia/admin/recruit/domain/RecruitDomain.java
@@ -22,7 +22,7 @@ public class RecruitDomain extends BaseDomain{
 	private String colTest2;
 	private String colTest3;
 	
-	private long atchNo;
+	private int atchNo;
 	
 	
 }
